move to formatUtils, return DiscourseNode
